+++ src/helpers.js
@@ -96,15 +96,23 @@ export function graphQLFetcher(graphQLParams) {
     });
   }
 //TODO readability
+
   export function parseQueryArray(queryArray) {
+    if(queryArray.length === 0) return;
     let queryString = "";
+    let newQueryArray;
     const wrap = (string) => "{" + string + "}";
-    (queryArray[queryArray.length - 1].length === 0) ? (queryString = wrap("id")) : (queryString = wrap(queryArray[queryArray.length - 1].join(" ")));
-    let newQueryArray = queryArray.slice(0, queryArray.length - 1);
+    if (Array.isArray(queryArray[queryArray.length - 1])){
+      (queryArray[queryArray.length - 1].length === 0) ? (queryString = wrap("id")) : (queryString = wrap(queryArray[queryArray.length - 1].join(" ")));
+      newQueryArray = queryArray.slice(0, queryArray.length -1);
+    } else {
+     queryString = "{FIELDS}";
+     newQueryArray = queryArray; 
+    }
     forEachRight(newQueryArray, (current) => {
       (Array.isArray(current)) ? (queryString = wrap(current.join(" ") + queryString)) : (queryString = wrap(current + queryString));
     });
-    return (queryString)
+    return queryString;
   }
